src/requirements.txt
boto3
strands-agents
yfinance